I fixed numerous compilation errors and applied structural improvements.
This addresses a significant number of compilation errors I identified in the initial `errors.xml` and subsequent build attempts.

Key changes include:

1.  **`BrushPersistence.cpp`**: I investigated C2451 and assumed it was a false positive or previously fixed.
2.  **`DeselectCommand.cpp`**: I added missing methods (`getSelectedPositions`, `deselectPosition`, `clearSelection`, `selectPosition`) to `Map.h` and `Map.cpp`.
3.  **`CreatureSpriteManager.h/cpp`**:
    *   I resolved undefined `Outfit` type errors by including `Outfit.h`.
    *   I corrected the `CreatureBrush::getOutfit()` return type.
    *   I addressed `QString::arg` issues by casting `Outfit` members.
    *   The C2659 error (`function as left operand`) was not found.
4.  **`DoodadBrush.cpp`**: I corrected enum names (`Brush::Type::Doodad`, `Brush::BrushShape::Square`).
5.  **`DoorBrush.cpp`**: I added a `MapPos(const QPointF&)` constructor and `Tile::getItems()` methods.
6.  **Iterative Build/Fix Cycle Progress**:
    *   `Item.h`: I removed the duplicate `Item::AttrDuration`.
    *   `ui/TileLockActions.h`: I added `#include <QShortcut>`.
    *   `ClipboardData.cpp`: I used `std::numeric_limits` to fix an overflow.
    *   `EditorSpriteIds.h`: I added `#include <QString>`.
    *   `EnhancedMapRenderer.h`: I added `#include <QHash>` and `#include <QString>`.
    *   `EditorSprite.cpp`: I fixed the `return &m_pixmaps.value()` logic.
    *   `EnhancedLightingSystem.h`: I made substantial updates including new enums, struct members, and method declarations to align with usage in the `.cpp` file.
    *   `EnhancedLightingSystem.cpp`: I applied initial fixes (constructors, `getEffectiveIntensity`, `updateBoundingRect`).

**Progress & Roadblocks:**
My original goal was to fix all compilation errors. While I resolved many, the project does not yet build successfully.
The primary roadblock I encountered was in `EnhancedLightingSystem.cpp`. Despite numerous attempts with increasingly focused efforts, applying further necessary changes (like refactoring color conversion utilities to static, renaming methods to match header declarations, and implementing stub functions) consistently failed. This suggests a persistent issue when modifying this specific complex file.

You expressed a desire to continue the iterative build-and-fix cycle until the build is successful. However, due to the issues with `EnhancedLightingSystem.cpp` and this being the designated final turn, I am presenting the progress made so far.

Further work would require resolving the modification issues for `EnhancedLightingSystem.cpp` or using an alternative strategy for that file, then continuing to fix errors in it and `EnhancedMapRenderer.h/cpp`, and iteratively building until all errors are cleared.

+++ Project_QT/src/CreatureSpriteManager.cpp
@@ -24,10 +24,10 @@ QString CreatureSpriteManager::generateCacheKey(int looktype, const Outfit* outf
                       .arg(height);
     if (outfit_ptr) { 
         key += QString("_head:%1_body:%2_legs:%3_feet:%4")
-                   .arg(outfit_ptr->lookHead)
-                   .arg(outfit_ptr->lookBody)
-                   .arg(outfit_ptr->lookLegs)
-                   .arg(outfit_ptr->lookFeet);
+                   .arg(static_cast<int>(outfit_ptr->lookHead))
+                   .arg(static_cast<int>(outfit_ptr->lookBody))
+                   .arg(static_cast<int>(outfit_ptr->lookLegs))
+                   .arg(static_cast<int>(outfit_ptr->lookFeet));
     }
     return key;
 }
